+++ test/server/i18next.functions.spec.js
@@ -33,6 +33,7 @@ describe('i18next.functions', function() {
       interpolationPrefix: '__',
       interpolationSuffix: '__',
       postProcess: '',
+      parseMissingKey: '',
       debug: false
     };