+++ test/server/i18next.init.spec.js
@@ -35,6 +35,7 @@ describe('i18next.init', function() {
       interpolationPrefix: '__',
       interpolationSuffix: '__',
       postProcess: '',
+      parseMissingKey: '',
       debug: false
     };